# docs/swagger.yaml
basePath: /api
definitions:
  common.APIResponse:
    description: 所有 API 响应的统一包装格式，包含状态码、消息、数据和 timestamp
    properties:
      code:
        description: 状态码，200 表示成功，其他表示错误
        example: 200
        type: integer
      data:
        description: 响应数据，成功时返回业务数据，错误时为 null
      message:
        description: 消息提示，成功时为"success"，错误时为错误描述
        example: success
        type: string
      timestamp:
        description: Unix 时间戳（秒）
        example: 1717200000
        type: integer
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
  models.ContentItem:
    description: 存储页面内容项，支持多语言和图标
    properties:
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        description: 描述（向后兼容）
        example: Description
        type: string
      enDescription:
        description: 英文描述
        example: We have a professional technical team
        type: string
      enTitle:
        description: 英文标题
        example: Professional Team
        type: string
      icon:
        description: 图标类名
        example: fa-users
        type: string
      id:
        type: integer
      imagePath:
        description: 图片路径
        example: /uploads/images/team.jpg
        type: string
      section:
        description: 所属区域/板块
        example: advantages
        type: string
      sortOrder:
        description: 排序顺序
        example: 1
        type: integer
      title:
        description: 标题（向后兼容）
        example: Title
        type: string
      updatedAt:
        type: string
      zhDescription:
        description: 中文描述
        example: 我们拥有专业的技术团队
        type: string
      zhTitle:
        description: 中文标题
        example: 专业团队
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    医疗器械内容管理系统 API 接口文档
    提供内容管理、模块配置、图片管理、多语言支持等功能
  title: Shirley CMS API
  version: "1.0"
paths:
...
  /api/admin/content:
    get:
      consumes:
      - application/json
      description: 获取所有或指定区域的内容项列表
      parameters:
      - description: 区域名称，不传则返回所有内容项
        in: query
        name: section
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 内容项列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ContentItem'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 获取内容项列表
      tags:
      - 内容管理
    post:
      consumes:
      - multipart/form-data
      description: 创建新的内容项，支持图片上传和多语言
      parameters:
      - description: 区域名称
        in: formData
        name: section
        required: true
        type: string
      - description: 中文标题
        in: formData
        name: zhTitle
        type: string
      - description: 英文标题
        in: formData
        name: enTitle
        type: string
      - description: 中文描述
        in: formData
        name: zhDescription
        type: string
      - description: 英文描述
        in: formData
        name: enDescription
        type: string
      - description: 图片文件
        in: formData
        name: image
        type: file
      - description: 排序顺序
        in: formData
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ContentItem'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 创建内容项
      tags:
      - 内容管理
  /api/admin/content/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的内容项
      parameters:
      - description: 内容项 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/common.APIResponse'
        "404":
          description: 内容项不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 删除内容项
      tags:
      - 内容管理
    put:
      consumes:
      - multipart/form-data
      description: 更新指定内容项的信息，支持图片上传和多语言
      parameters:
      - description: 内容项 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 区域名称
        in: formData
        name: section
        type: string
      - description: 中文标题
        in: formData
        name: zhTitle
        type: string
      - description: 英文标题
        in: formData
        name: enTitle
        type: string
      - description: 中文描述
        in: formData
        name: zhDescription
        type: string
      - description: 英文描述
        in: formData
        name: enDescription
        type: string
      - description: 图片文件
        in: formData
        name: image
        type: file
      - description: 排序顺序
        in: formData
        name: sort_order
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ContentItem'
              type: object
        "404":
          description: 内容项不存在
          schema:
            $ref: '#/definitions/common.APIResponse'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      security:
      - APIKey: []
      summary: 更新内容项
      tags:
      - 内容管理
...
  /api/public/content:
    get:
      consumes:
      - application/json
      description: 获取所有或指定区域的公开内容项（无需认证）
      parameters:
      - description: 区域名称，不传则返回所有内容项
        in: query
        name: section
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 内容项列表
          schema:
            allOf:
            - $ref: '#/definitions/common.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ContentItem'
                  type: array
              type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/common.APIResponse'
      summary: 获取公开内容项
      tags:
      - 公开接口
...
securityDefinitions:
  APIKey:
    description: JWT Token 认证，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"